hydra dispatching: move to dispatch_hydra.rs
